Add initial work for colorization.
The interface is a little wonky because of how intensely generic the image crate is, but
it should be good enough. It allows people to define their own colorizers and implement
a simple transform from an input pixel to an RGB pixel (or RGBA if desired), and get the
rest of the behaviors for free.

Of course, given that I'm still working on this and haven't yet begun to use this interface,
things may still change. If it happens, it happens; this is early versions yet.